Move packager into a package
