+++ cvra_packager/CMakeLists.txt.jinja
