+++ cvra_packager/Makefile.arm.jinja
