+++ cvra_packager/Makefile.base.jinja
