+++ cvra_packager/__init__.py
@@ -1,0 +1,1 @@
+from .packager import *
